requirements.txt
yfinance>=0.2.36
pandas>=2.1.4
numpy>=1.26.2
requests>=2.31.0
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
python-dotenv>=1.0.1
